pyproject.toml
[project]
name = "gsa-certificate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
    "pymupdf>=1.23.0",
    "pillow>=10.0.0",
]
